refactor : FarePolicyTemplate 패키지 변경 및 메서드 네이밍 변경
